[refactor] dto를 response, request로 분리 #67
