+++ src/main/java/dev/book/global/sse/dto/response/SseAchievementResponse.java
@@ -1,4 +1,4 @@
-package dev.book.global.sse.dto;
+package dev.book.global.sse.dto.response;
 
 public record SseAchievementResponse(String lastEventId, String title, String content, String name){
 }
